src/Python/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
opencv-python-headless
numpy
pillow
google-genai
firebase-admin
